% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{columns}
\title{The \code{columns} method for \code{SomaDb} objects}
\value{
A character vector of all available database columns.
}
\examples{
# Retrieve all database columns
columns(SomaScan.db)
}
\author{
Amanda Hiser
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AnnotationDbi}{\code{\link[AnnotationDbi:AnnotationDb-class]{columns}}}
}}

