% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{SomaDb-class}
\alias{SomaDb-class}
\alias{.SomaDb}
\title{The SomaDb S4 class}
\description{
The SomaDb class is a subclass of the ChipDb (and, by extension,
the AnnotationDb) class defined in the AnnotationDbi package. It is designed
to hold annotations for SomaScan data.
}

